requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp_socks==0.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bottle==0.12.25
Brotli==1.1.0
browser-cookie3==0.19.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
curl_cffi==0.7.1
defusedxml==0.7.1
dnspython==2.6.1
duckduckgo_search==6.2.5
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
g4f==0.3.2.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
outcome==1.3.0.post0
pillow==10.4.0
platformdirs==4.2.2
plyer==2.1.0
primp==0.5.5
proxy_tools==0.1.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyExecJS==1.5.1
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
python-socks==2.5.0
pythonnet==3.0.3
pywebview==5.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
selenium==4.23.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
tinycss2==1.3.0
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.2
uvicorn==0.30.3
watchfiles==0.22.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.4
